BOMB is BOMBING!
added the bomb explosion functionality.
if a bomb hits an object it wipes every block in 3X3 area within the object.
the explosion is a function of Board class.
Explosion function checks every possible position in which the bomb can land and it makes the changes accordingly. (order of erasing will be uploaded).

the function also updates the values in the game matrix and in the indicators matrix.

:+1:  @lermanori